Redistribución de los ficheros tras las correcciones y sugerencias del profesor
