+++ funcionesV2.py
@@ -1,0 +1,8 @@
+from funcionesV1 import *
+
+
+def detectar_cochesV2():
+    return
+
+def contar_cochesV2():
+    return
